requirements.txt
pillow
icontract
python-telegram-bot
psycopg[binary]
